Check for "null" strings in apikey and replace with null
When saving the settings, a null apikey input will return "null" from the jquery .val() call. Wrapped call in a "null" check to properly store this in the script storage as a null. Also wrapped the settings init with the same check to fix any previously stored "null" strings from users of the script